Show response_timestamp in Invoice Rejection Detail window

+++ backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_440_response/src/main/java-xjc/de/metas/vertical/healthcare_ch/forum_datenaustausch_ch/invoice_440/response/PayloadType.java
@@ -84,7 +84,7 @@ public class PayloadType {
     protected Boolean creditAdvice;
     @XmlAttribute(name = "response_timestamp", required = true)
     @XmlSchemaType(name = "unsignedLong")
-    protected BigInteger responseTimestamp;
+    protected int responseTimestamp;
 
     /**
      * Gets the value of the credit property.
@@ -323,10 +323,10 @@ public class PayloadType {
      * 
      * @return
      *     possible object is
-     *     {@link BigInteger }
+     *     {@link int }
      *     
      */
-    public BigInteger getResponseTimestamp() {
+    public int getResponseTimestamp() {
         return responseTimestamp;
     }
 
@@ -335,10 +335,10 @@ public class PayloadType {
      * 
      * @param value
      *     allowed object is
-     *     {@link BigInteger }
+     *     {@link int }
      *     
      */
-    public void setResponseTimestamp(BigInteger value) {
+    public void setResponseTimestamp(int value) {
         this.responseTimestamp = value;
     }
