+++ backend/vertical-healthcare_ch/forum_datenaustausch_ch.invoice_xversion/src/main/java/de/metas/vertical/healthcare_ch/forum_datenaustausch_ch/invoice_xversion/response/model/XmlPayload.java
@@ -36,7 +36,7 @@ public class XmlPayload
 {
 
 	@NonNull
-	BigInteger responseTimestamp;
+	int responseTimestamp;
 
 	@NonNull
 	XmlInvoice invoice;
